[codegen][compiler] Convert usages of FixedArray::length to ulength
Improve robustness in the sandbox attacker model by using unsigned
integer length for FixedArray (see: crrev.com/c/7544875),
length(kAcquireLoad) now returns SafeHeapObjectSize. Additionally,
convert index and count values in HandlerTable to use uint32_t.

Bug: 375937549
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7567310
Commit-Queue: Arash Kazemi <arashk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#105288}

+++ src/execution/isolate.cc
@@ -3043,7 +3043,7 @@ HandlerTable::CatchPrediction PredictExceptionFromBytecode(
   HandlerTable table(bytecode);
   int handler_index = table.LookupHandlerIndexForRange(code_offset);
   if (handler_index < 0) return HandlerTable::UNCAUGHT;
-  return table.GetRangePrediction(handler_index);
+  return table.GetRangePrediction(static_cast<uint32_t>(handler_index));
 }
 
 HandlerTable::CatchPrediction PredictException(const FrameSummary& summary,
